Polish the code after separating DLedgerProxy module
