+++ command/src/main/java/io/openmessaging/storage/dledger/command/AppendCommand.java
@@ -14,21 +14,18 @@
  * limitations under the License.
  */
 
-package io.openmessaging.storage.dledger.cmdline;
+package io.openmessaging.storage.dledger.command;
 
 import com.alibaba.fastjson.JSON;
 import com.beust.jcommander.Parameter;
-import com.beust.jcommander.Parameters;
 import io.openmessaging.storage.dledger.client.DLedgerClient;
-import io.openmessaging.storage.dledger.entry.DLedgerEntry;
-import io.openmessaging.storage.dledger.protocol.GetEntriesResponse;
+import io.openmessaging.storage.dledger.protocol.AppendEntryResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-@Parameters(commandDescription = "Get data from DLedger server")
-public class GetCommand extends BaseCommand {
+public class AppendCommand extends BaseCommand {
 
-    private static Logger logger = LoggerFactory.getLogger(GetCommand.class);
+    private static Logger logger = LoggerFactory.getLogger(AppendCommand.class);
 
     @Parameter(names = {"--group", "-g"}, description = "Group of this server")
     private String group = "default";
@@ -36,19 +33,19 @@ public class GetCommand extends BaseCommand {
     @Parameter(names = {"--peers", "-p"}, description = "Peer info of this server")
     private String peers = "n0-localhost:20911";
 
-    @Parameter(names = {"--index", "-i"}, description = "get entry from index")
-    private long index = 0;
+    @Parameter(names = {"--data", "-d"}, description = "the data to append")
+    private String data = "Hello";
+
+    @Parameter(names = {"--count", "-c"}, description = "append several times")
+    private int count = 1;
 
     @Override
     public void doCommand() {
         DLedgerClient dLedgerClient = new DLedgerClient(group, peers);
         dLedgerClient.startup();
-        GetEntriesResponse response = dLedgerClient.get(index);
-        logger.info("Get Result:{}", JSON.toJSONString(response));
-        if (response.getEntries() != null && response.getEntries().size() > 0) {
-            for (DLedgerEntry entry : response.getEntries()) {
-                logger.info("Get Result index:{} {}", entry.getIndex(), new String(entry.getBody()));
-            }
+        for (int i = 0; i < count; i++) {
+            AppendEntryResponse response = dLedgerClient.append(data.getBytes());
+            logger.info("Append Result:{}", JSON.toJSONString(response));
         }
         dLedgerClient.shutdown();
     }
