+++ command/src/main/java/io/openmessaging/storage/dledger/command/ReadFileCommand.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.openmessaging.storage.dledger.cmdline;
+package io.openmessaging.storage.dledger.command;
 
 import com.alibaba.fastjson.JSON;
 import com.beust.jcommander.Parameter;
