+++ proxy/src/test/java/io/openmessaging/storage/dledger/proxy/ProxyAppendAndGetTest.java
@@ -14,19 +14,16 @@
  * limitations under the License.
  */
 
-package io.openmessaging.storage.dledger;
+package io.openmessaging.storage.dledger.proxy;
 
 import io.openmessaging.storage.dledger.DLedgerConfig;
 import io.openmessaging.storage.dledger.DLedgerServer;
-import io.openmessaging.storage.dledger.ServerTestHarness;
 import io.openmessaging.storage.dledger.client.DLedgerClient;
 import io.openmessaging.storage.dledger.entry.DLedgerEntry;
 import io.openmessaging.storage.dledger.protocol.AppendEntryRequest;
 import io.openmessaging.storage.dledger.protocol.AppendEntryResponse;
 import io.openmessaging.storage.dledger.protocol.DLedgerResponseCode;
 import io.openmessaging.storage.dledger.protocol.GetEntriesResponse;
-import io.openmessaging.storage.dledger.proxy.DLedgerProxy;
-import io.openmessaging.storage.dledger.proxy.DLedgerProxyConfig;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
