+++ proxy/src/test/java/io/openmessaging/storage/dledger/proxy/ServerTestBase.java
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-package io.openmessaging.storage.dledger;
-import io.openmessaging.storage.dledger.util.FileTestUtil;
+package io.openmessaging.storage.dledger.proxy;
+import io.openmessaging.storage.dledger.proxy.util.FileTestUtil;
 
 import java.util.ArrayList;
 import java.util.List;
